Tests - Move RME test base class to corresponding RME folder.
