+++ tests/RME/__init__.py
@@ -1,0 +1,5 @@
+from .RME_test_case import RMETestCase
+
+__all__ = [
+    RMETestCase
+]
