+++ tests/data/hrrr/test_hrrr_gold.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pandas as pd
 
-from tests.RME_test_case import RMETestCase
+from tests.RME import RMETestCase
 from weather_forecast_retrieval.data.hrrr import FileLoader
