+++ tests/data/hrrr/test_http_retrieval.py
@@ -3,7 +3,7 @@ from datetime import datetime, timedelta
 
 import pandas as pd
 
-from tests.RME_test_case import RMETestCase
+from tests.RME import RMETestCase
 from tests.helpers import skip_external_http_request
 from weather_forecast_retrieval.data.hrrr import HttpRetrieval
