+++ tests/test_hrrr_archive.py
@@ -7,7 +7,7 @@ import unittest
 import numpy as np
 import pandas as pd
 
-from tests.RME_test_case import RMETestCase
+from tests.RME import RMETestCase
 from tests.helpers import skip_external_http_request
 from weather_forecast_retrieval import hrrr_archive
